Add string util to camel case project names that contains hyphens.
The reason to fix this is because npm does not seem to function properly
when the project name specified is containing hyphens. The idea is that
it should not be necessary for a user to think about this, but this
ought to be handled so the user's focus can be on implementing awesome
things.